requirements.txt
scikit-learn
joblib
pillow
